fix: addressed PR comments
